feat: userReader 수정
