+++ src/main/java/com/example/surveyapp/domain/survey/domain/strategy/MultiChoiceAnswerStrategy.java
