+++ src/main/java/com/example/surveyapp/global/reader/UserReader.java
@@ -1,9 +1,8 @@
 package com.example.surveyapp.global.reader;
 
-import com.example.surveyapp.domain.user.domain.model.User;
 import com.example.surveyapp.domain.user.domain.model.UserRoleEnum;
 
 public interface UserReader {
     void validateUserIdOrThrow(Long userId);
-    boolean isSurveyor(Long userId);
+    boolean validateUserRole(Long userId, UserRoleEnum userRole);
 }
